Split NDJSON row driver
